Do not copy to system clipboard by default
Often I copy something from the browser to the system clipboard and then
switch to lvim and use the change command with a text object to
overwrite some text. Using the change command overwrites the system
clipboard which results in the loss of the previously copied text. This
is annoying the least said.