Fix configuration for the better_whitespace plugin
A setup function is now responsible for setting up the plugin in terms
of global configuration parameters and autocommands. All relevant code
is now in module user.plugins.better_whitespace.
The autocommand to disable stripping whitespace for markdown is now also
moved there and wrapped in a plugin specific autocommand group.